Move navbar into common components
